[REFACTOR]  Split filter classes into separate files
